backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-groq
langchain-community
duckduckgo-search
supabase
edge-tts
gTTS
requests
pydantic